requirements.txt
httpx==0.27.0
python-dotenv==1.0.1
aiogram==3.13.1
aiofiles==24.1.0
aiosqlite==0.20.0
fastapi==0.115.0
uvicorn[standard]==0.32.0
jinja2==3.1.4

